+++ src/cli/sourcecred.js
@@ -8,6 +8,7 @@ import {VERSION_SHORT} from "../core/version";
 import help from "./help";
 import load from "./load";
 import analyze from "./analyze";
+import exportGraph from "./exportGraph";
 
 const sourcecred: Command = async (args, std) => {
   if (args.length === 0) {
@@ -25,6 +26,8 @@ const sourcecred: Command = async (args, std) => {
       return load(args.slice(1), std);
     case "analyze":
       return analyze(args.slice(1), std);
+    case "export-graph":
+      return exportGraph(args.slice(1), std);
     default:
       std.err("fatal: unknown command: " + JSON.stringify(args[0]));
       std.err("fatal: run 'sourcecred help' for commands and usage");