+++ src/core/repoIdRegistry.test.js
@@ -3,12 +3,19 @@
 import {
   toJSON,
   fromJSON,
-  addRepoId,
+  addEntry,
+  getEntry,
   emptyRegistry,
   type RepoIdRegistry,
+  getRegistry,
+  writeRegistry,
+  REPO_ID_REGISTRY_FILE,
 } from "./repoIdRegistry";
 
 import {makeRepoId} from "../core/repoId";
+import tmp from "tmp";
+import path from "path";
+import fs from "fs";
 
 describe("core/repoIdRegistry", () => {
   describe("fromJSON compose on", () => {
@@ -26,15 +33,16 @@ describe("core/repoIdRegistry", () => {
       ]);
     });
   });
-  describe("addRepoId", () => {
+
+  describe("addEntry", () => {
     it("adds to empty registry", () => {
       expect(
-        addRepoId(emptyRegistry(), {repoId: makeRepoId("foo", "bar")})
+        addEntry(emptyRegistry(), {repoId: makeRepoId("foo", "bar")})
       ).toEqual([{repoId: makeRepoId("foo", "bar")}]);
     });
     it("adds to nonempty registry", () => {
       const registry = [{repoId: makeRepoId("foo", "bar")}];
-      expect(addRepoId(registry, {repoId: makeRepoId("zoo", "zod")})).toEqual([
+      expect(addEntry(registry, {repoId: makeRepoId("zoo", "zod")})).toEqual([
         {repoId: makeRepoId("foo", "bar")},
         {repoId: makeRepoId("zoo", "zod")},
       ]);
@@ -44,7 +52,7 @@ describe("core/repoIdRegistry", () => {
         {repoId: makeRepoId("zoo", "zod")},
         {repoId: makeRepoId("foo", "bar")},
       ];
-      expect(addRepoId(registry, {repoId: makeRepoId("foo", "bar")})).toEqual(
+      expect(addEntry(registry, {repoId: makeRepoId("foo", "bar")})).toEqual(
         registry
       );
     });
@@ -53,13 +61,70 @@ describe("core/repoIdRegistry", () => {
         {repoId: makeRepoId("zoo", "zod")},
         {repoId: makeRepoId("foo", "bar")},
       ];
-      expect(addRepoId(registry, {repoId: makeRepoId("zoo", "zod")})).toEqual([
+      expect(addEntry(registry, {repoId: makeRepoId("zoo", "zod")})).toEqual([
         {repoId: makeRepoId("foo", "bar")},
         {repoId: makeRepoId("zoo", "zod")},
       ]);
     });
   });
+
+  describe("getEntry", () => {
+    it("returns the matching entry by RepoId", () => {
+      const entry = {repoId: makeRepoId("zoo", "zod")};
+      const registry = addEntry(emptyRegistry(), entry);
+      expect(getEntry(registry, entry.repoId)).toBe(entry);
+    });
+    it("returns undefined if there is no matching entry", () => {
+      expect(getEntry(emptyRegistry(), makeRepoId("foo", "bar"))).toBe(
+        undefined
+      );
+    });
+  });
+
   it("empty registry is empty", () => {
     expect(emptyRegistry()).toEqual([]);
+  });
+
+  describe("{get,write}Registry", () => {
+    const repoId = () => makeRepoId("foo", "bar");
+    const entry = () => ({repoId: repoId()});
+    const registry = () => addEntry(emptyRegistry(), entry());
+
+    it("getRegistry returns empty registry if nothing present", () => {
+      const dirname = tmp.dirSync().name;
+      expect(getRegistry(dirname)).toEqual(emptyRegistry());
+    });
+
+    it("writeRegistry writes a repoIdRegistry to the directory", () => {
+      const dirname = tmp.dirSync().name;
+      const registryFile = path.join(dirname, REPO_ID_REGISTRY_FILE);
+
+      expect(fs.existsSync(registryFile)).toBe(false);
+      writeRegistry(registry(), dirname);
+      expect(fs.existsSync(registryFile)).toBe(true);
+
+      const contents = fs.readFileSync(registryFile);
+      const registryJSON = JSON.parse(contents.toString());
+      expect(toJSON(registry())).toEqual(registryJSON);
+    });
+
+    it("getRegistry returns the registry written by writeRegistry", () => {
+      const dirname = tmp.dirSync().name;
+      const repoId = makeRepoId("foo", "bar");
+      const entry = {repoId};
+      const registry = addEntry(emptyRegistry(), entry);
+      writeRegistry(registry, dirname);
+      expect(getRegistry(dirname)).toEqual(registry);
+    });
+
+    it("writeRegistry overwrites any existing registry", () => {
+      const dirname = tmp.dirSync().name;
+      const repoId = makeRepoId("foo", "bar");
+      const entry = {repoId};
+      const registry = addEntry(emptyRegistry(), entry);
+      writeRegistry(registry, dirname);
+      writeRegistry(emptyRegistry(), dirname);
+      expect(getRegistry(dirname)).toEqual(emptyRegistry());
+    });
   });
 });