+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## [Unreleased]
-
 - Add description tooltips for node and edge types in the weight configuration UI (#1081)
+- Add the `export-graph` command (#1110)
 - Enable loading private repositories (#1085)
 - Enable setting type weights to 0 in the UI (#1005)
 - Add support for 🚀 and 👀 reaction types (#1068)